Fix SDL2-example for Linux
The header is called "SDL_opengl.h" - case matters on Linux.

The main() function must indeed be called main(), on Windows it'll be
renamed to SDL_main (for SDL2main.lib) by #define in SDL_main.h
(which is included by SDL.h)

I added an entry to the example's README that describes how to build
it on Linux.