supress potentially unused identifier
we pull the lib in as a dep, but our toolchains are configured to error on warning... 
you don't have to accept this.  Just throwing it out there.  Maybe I'll just pragma suppress the warning in a wrapper header around this library or something instead (a wrapper in our own codebase that is).