Examples: Metal: Restore state after UserCallback
In other backends it is possible for the UserCallback to read the current state, do whatever state changes it needs, and then restore the imgui state. In Metal there doesn't seem to be a way for a UserCallback to save the existing state before changing it so we should restore our state after the UserCallback is complete.