Fix filter-flags disabling newline in multiline inputs
Implements #4409: `ImGui::InputTextMultiLine` stops being multi-line when the flags also
specify one of the named character filters.

```cpp
    char b1[256] = "1A2B\n3C4D\n", b2[256] = "1A2B\n3C4D";
    ImGui::InputTextMultiline("##t1", b1, 255, ImVec2(0,0), ImGuiInputTextFlags_None);
    ImGui::InputTextMultiline("##t2", b2, 255, ImVec2(0,0), ImGuiInputTextFlags_CharsHexadecimal);
```

In the above example, it is not possible to add a new newline to the
input box, because that input gets filtered out.

This fix short-circuits the existing checks on newline and tab so that
are not subsequently nixed by a named filter.