Fixed cropping oversight in the allegro implementation
I noticed when I put a window near the bottom right of the display then resizing the window to be small, that anything inside started to be shown outside of it. This is more obvious with large images.

I tracked down the problem, and it was that al_set_clipping_rectangle takes position and size, not minimum and maximum position.