Avoiding using SDL_GL_GetDrawableSize
This is (kind of) an OpenGL-only function, which should be avoided
when SDL2 isn't using OpenGL.

The only alternative that is recommended is
SDL_GetRendererOutputSize, which limits this fix to the SDL_Renderer
backend. Still, I think it's better than nothing.

I say that SDL_GL_GetDrawableSize is "kind of" OpenGL-only because it
does technically work even when SDL2 isn't using OpenGL. It's just
that it becomes a shim to SDL_GetWindowSize, which is not suitable
for high-DPI usage because it reflects the size of the window in
screen coordinates, not actual pixels, so it really should be avoided
when not using OpenGL.