Move test from ImGui_ImplVulkan_RenderDrawData to ImGui_ImplVulkan_CreateFontsTexture
The test should never fail: if it fails then the user is on
a 32-bit machine and neglected to redefine ImTextureID.

It confused me a lot that this test was done where it is done
(in code executed every frame). Although it should end up as a
no-op, it suggests that the user is *allowed* to redefine
ImTextureID to something that is less than 64-bit, in which
case the `desc_set[0] = bd->FontDescriptorSet;` would be
necessary, under certain circumstances. And that suggests
that ImTextureID is *still* under the users control, despite
using this backed.

This backend (imgui_impl_vulkan.cpp) uses ImTextureID to store
a VkDescriptorSet - no matter what the user does. And therefore
this backend should not allow the user to have an ImTextureID
that can't contain 64-bits of data, at all.

The patch moves the test of a 32-bit user neglecting to redefine
ImTextureID to a more logical place and simply always fails if
they didn't, as it should with this backend when ImTextureID can't
hold a VkDescriptorSet.