Fix #4947
Implement a function to update the undo/redo state for an `InputText`
after a user callback has modified the buffer.

This function is a bit weird in that it uses the wide string from the
state for the old text and a utf8 buffer for the new text. It does this
because these are the formats that are already available at the time we
need them without requiring additional storage.