Remove DebugAllocHook from MemAlloc and MemFree
--
This commit is a preparation toward adding ImGui apis with explicit context
and making ImGui applications being able to use multiple context at the same time
whatever their concurrency model.
--

DebugAllocHook forces MemAlloc and MemFree to be dependent on ImGuiContext.
In the scenario where the context need to be explicit, it means forward the
context to every MemAlloc and MemFree calls. This sounds cumbersome.

Instead, this commit remove the use of DebugAllocHook to make MemAlloc and
MemFree independent from ImGuiContext.