Make MetricsActiveAllocations thread-safe and independent from GImGui context
--
This commit is a preparation toward adding ImGui apis with explicit context
and making ImGui applications being able to use multiple context at the same time
whatever their concurrency model.
--

Even in a scenario where multiple ImGuiContext are active at the same time
ImGui::MemAlloc and ImGui::MemFree will still be global and context independent.

The implementation of ImGui::MemAlloc and ImGui::MemFree should not depend
on the ImGuiContext.

Prior to this commit, the implementation was depending on ImGuiContext to store
active allocation metrics. Instead this commit makes this metrics be an atomic
counter defined next to GImAllocatorAllocFunc and GImAllocatorUserData globals