Make win32 clipboard text and platform ime functions not depend on GImGui
--
This commit is a preparation toward adding ImGui apis with explicit context
and making ImGui applications being able to use multiple context at the same time
whatever their concurrency model.
--

Prior to this commit `GetClipboardTextFn_DefaultImpl` on windows was depending
on the global GImGui context.
Instead this commit is forwarding the ImGuiContext as user data.

Prior to this commit `SetPlatformImeDataFn_DefaultImpl` on windows was depending
on the global GImGui context through a call to ImGui::GetIO().ImeWindowHandle.
The call has been remove, and ImeWindowHandle is forwarded by setting
ImGuiViewport::PlatformHandleRaw to ImeWindowHandle when PlatformHandleRaw
is null.