Skip %+ and %# format flags for scanning
Fix typed in values are ignored when SliderInt's format specifier
includes the + flag:

> +: the sign of signed conversions is always prepended to the result of
> the conversion (by default the result is preceded by minus only when it
> is negative)

Also skip the # flag since it seems valid to use (esp. for octal and hex):

> # : alternative form of the conversion is performed. See the table
> below for exact effects otherwise the behavior is undefined.

There are two additional unhandled flags that only affect padding:
- and space. Formatting flags don't make sense in a SliderInt's format
string, so I've omitted them.

Test
In this example, editing the value by dragging it to +80º works but when
Ctrl-clicking and typing in 10 it reverts to the previous value after
hitting Enter:
	if (ImGui::Begin("Hello"))
	{
		static int val = 0;
		ImGui::SliderInt("Slider", &val, 0, 100, "%+dº");
		ImGui::End();
	}