Backends: Vulkan: Create a custom pipeline for viewports
The main render pass may have a different configuration than the viewport window render pass.
Vulkan can reuse the pipeline between different render passes only if the render pass configuration is the same.
To fix it, when it's time to RenderDrawData for the viewport window and a pipeline is not yet available, a new pipeline is created for that window to be sure that is compatible with the render pass.