Fix SDL Gamepad handling
SDL_OpenGamepad(0) doesn't work with latest SDL3.
Now we need to get all the gamepags, then we need to open it:
    int pads = 0;
    auto ids = SDL_GetGamepads(&pads);
    auto joyId = SDL_OpenGamepad(ids[0]);
    ...
    SDL_CloseGamepad(joyId);

A new function ImGui_ImplSDL3_SetGamepad is needed to allow the user to
set which gamepad will be used for Ui Navigation, as we shouldn't open/
close the gamepad for each frame.