Update imgui_impl_vulk backend to define NOMINMAX when VK_USE_PLATFORM_WIN32_KHR is defined
`#include <vulkan/vulkan.h>` includes `windows.h` if `VK_USE_PLATFORM_WIN32_KHR` is defined. `windows.h` defines `min` and `max` macros if `NOMINMAX` is not defined. This change now defines it if it is not defined then undefines it if it wasn't defined. This should avoid side effects for builds which both do and don't define `NOMINMAX`.