Examples: SDL3: change to using $(var) syntax for environment paths
If you upgrade the project to the latest vs2022  it fails to find the SDL3_DIR properly. This changes it to use a more modern syntax that works in both, and finds the correct paths without issue.

e.g. they only have v143, and latest win 10 sdk so are forced to upgrade from vs2015
Before: slightly obscure errors, and confusion
After: everything "just works"