Added support to render OpenType SVG fonts using plutosvg as an alternative to lunasvg (fix #7187)
<SVG Fonts include a set of SVG documents. As per the [OpenType specification](https://learn.microsoft.com/en-us/typography/opentype/spec/svg#glyph-identifiers),
some SVG fonts (such as NotoColorEmoji) may group several glyphs in a common svg document (by selecting a subset of the elements in this document).

LunaSvg does support fonts where each glyph is associated to a distinct document. Unfortunately, it is not able to render
a subset of a svg document, and will likely not be able to do so in the future.

Its cousin project plutosvg (by the same author), is able to do it, and provides ready to use freetype hooks.

Example: https://github.com/sammycage/lunasvg/issues/150 shows an example where a single svg document
included in the font may contains thousands of glyphs (each glyph is a subset of the svg document).

Pros and Cons
-------------
Since this commit adds some complexity in the code here is a study of the pros and cons:

Pros:
- plutosvg is faster than lunasvg
- freetype hooks are provided by plutosvg (no need to provide them in imgui_freetype.cpp)

Cons:
- the compilation setup for plutosvg is a bit more complex than for lunasvg (requires plutovg + plutosvg)
- no offical release is available yet for plutosvg. No vcpkg package available yet.
- having two competing compilation flags is not ideal (IMGUI_ENABLE_FREETYPE_LUNASVG vs IMGUI_ENABLE_FREETYPE_PLUTOSVG)
  (it may be possible to remove IMGUI_ENABLE_FREETYPE_LUNASVG in the future, at the cost of breaking some users build upon upgrade)

Compilation hints for plutovg/plutosvg
--------------------------------------
_Compilation hints for plutovg_
- Compile all source files in `plutovg/source/*.c`
- Add include directory: `plutovg/include` + `plutovg/stb`

_Compilation hints for plutosvg_
- Compile `plutosvg/source/plutosvg.c`
- Add include directory: `plutosvg/source`
- Add define: `PLUTOSVG_HAS_FREETYPE`
- Link with: plutovg, freetype

Demonstration repository
-------------------------
https://github.com/pthom/lunasvg_perf_issue