Removed column skip items assert
because IsVisibleY was changed to always be true for columns and is_visible changed to only care about the X-axis, the assert for is_visible on if the column items should be skipped is now an erroneous assert because of the case where the column is part of a table that has clipped off the top or bottom of the cliprect.  In those cases, is_visible will always be true (horizontally) but hostSkipItems will be false.