Default to first monitor is viewport is outside bounds
Fixes https://github.com/ocornut/imgui/issues/8385 by returning a valid monitor from FindPlatformMonitorForRect. This function only checks the rect if more than one monitor is present, so seems reasonable to assume that it should return a valid monitor even if the rect doesn't clip with the monitor.

Issue was causing an assert when a viewport was loaded out of the bounds of any of the monitors. Before the assert was introduced (in d66f4e5890f8df2a52434aac2d14ff381663b1c1) the viewport would be eventually clamped with ClampWindowPos using g.FallbackMonitor.