Fix wgpu image binddgroups
On WebGPU backend, in order to display an image, an ImageBindGroup needs to be created for every new textureView. The backend creates and caches these imagebindgroups but it doesn't consider the fact that textureviews can be created/destroyed on the user side.

The webgpu backend can reuse the pointer for the textureview side, so when the user destroys and creates a new textureview, the cached imagebindgroup will try to use a destroyed resource and probably crash.

In order to fix it, we remove the imagebindgroups created during render on every frame.

This PR also updates the code to the latest webgpu-native, where some changes there were breaking the compilation