Change ImFontConfig::FontNo to ImS32 (was ImS8)
When used with FreeType this value is passed as `face_index` which is interpreted as two 16-bit values: the lower 16 bits are the "face index" and the upper 16 are the "named instance index" for the selected face. Specifying the instance index allows for things like predefined font weights (Bold, SemiBold, Light, etc.) in many fonts.