Added ImGui::FreeDemoWindow to allow manual releasing of demo heap allocations
I noticed that if I used the demo window when also using ImGui::SetAllocatorFunctions for allocation tracking there would still be around 250 live allocations (0.3MiB) when my allocator runs it's leak detection just before the end of my application.

After looking inside of ImGui::ShowDemoWindow I can see that we build a
static ImGuiDemoWindowData on the first call to the show the demo window, but this does mean there is no way to free these heap allocations manually, we must wait for application exit for the destructor to be called which then frees the demo tree data.

This change moves the local ImGuiDemoWindowData static out of ImGui::ShowDemoWindow and into the global scope of `imgui_demo.cpp` this allows for adding a manual demo free function called `FreeDemoWindow`, if this function is not manually called the behavior is the same as before and the debug data is freed on app exit.