Fixed an assert in background dimming code,  which could trigger after using gamepad/keyboard to move a window to another viewport. (#9053)
Due to missing ->Active check the dimming could render behind a window which was not submitted during the frame, appending to left-over ImDrawList data and in some situations tripping on the assert in RenderDimmedBackgroundBehindWindow().
The assumption that the PushClipRect won't match previous command is arguably a bit fragile, but this is what the assert is for.